split up watcher logic into helper
